requirements.txt
appdirs==1.4.3
Django==1.9
django-json-field==0.5.7
django-jsonfield==1.0.1
djangorestframework==3.6.2
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
six==1.10.0
